Fix bugs
Fixed bug where after pressing equals button, inputting more digits added onto
the previous second number instead of starting a new number. This was due to
order of if statements checking if an operator existed before checking if a
calculation result exists.
Fixed bug when first or second number are 0, user can't move on, or if the result
is equal to 0 then starting a new calculation doesn't work. This is due to
checking if a number if valid using parseFloat() to pass if statements instead of
checking if a number is a number.